Consolidating PlanetParams objects
