+++ src/eureka/S5_lightcurve_fitting/models/__init__.py
@@ -1,5 +1,5 @@
 from .Model import Model, CompositeModel
-from .AstroModel import AstroModel
+from .AstroModel import AstroModel, PlanetParams
 from .BatmanModels import BatmanTransitModel, BatmanEclipseModel
 from .CentroidModel import CentroidModel
 from .DampedOscillator import DampedOscillatorModel
